+++ plotoptix/__init__.py
@@ -12,8 +12,8 @@ __all__ = ["enums", "materials", "utils", "npoptix", "tkoptix"]
 
 __author__  = "Robert Sulej, R&D Team <dev@rnd.team>"
 __status__  = "beta"
-__version__ = "0.4.7"
-__date__    = "27 Aug 2019"
+__version__ = "0.4.8"
+__date__    = "7 Sep 2019"
 
 import logging