+++ plotoptix/materials.py
@@ -30,8 +30,19 @@ m_diffuse = {
       "VarInt": { "material_flags": 2 }
     }
 """
-Standard diffuse material. Note it is available by default under the name "diffuse".
+Lambertian diffuse material. Note it is available by default under the name "diffuse".
 Use color components range <0; 1>.
+"""
+
+m_matt_diffuse = {
+      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::diffuse_closest_hit" ],
+      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
+      "VarInt": { "material_flags": 2 },
+      "VarFloat": { "surface_roughness": 1 }
+    }
+"""
+Oren-Nayar diffuse material. Surface roughness range is <0; inf), 0 is equivalent to
+the Lambertian "diffuse" material.
 """
 
 m_mirror = {
@@ -43,32 +54,23 @@ m_mirror = {
       }
 }
 """
-100% reflective mirror, quite simple to calculate and therefore a fast material. Use
-surface_albedo (range <0; 1>) to colorize reflection.
+Reflective mirror, quite simple to calculate and therefore fast material. Note, this material
+has default values: ``reflectivity_index = 1`` and ``reflectivity_range = 1``. In this configuration
+the shading algorithm overrides ``surface_albedo`` with the color assigned to each primitive (RGB
+range <0; 1>), which results with colorized reflections.
 """
 
 m_metalic = {
       "ClosestHitPrograms": ["0::path_tracing_materials.ptx::reflective_closest_hit"],
       "AnyHitPrograms": ["1::path_tracing_materials.ptx::any_hit"],
       "VarInt": { "material_flags": 6 },
-      "VarFloat": {
-        "reflectivity_index": 0.95,
-        "reflectivity_range": 1.0,
-      },
-      "VarFloat3": {
-        "refraction_index": [ 2.5, 2.5, 2.5 ],
-        "surface_albedo": [ 1.0, 1.0, 1.0 ]
-      }
+      "VarFloat": { "surface_roughness": 0.002 },
 }
 """
-Strongly reflective, metalic material. Use surface_albedo (range <0; 1>) to colorize
-reflection. Standard color assigned to each primitive is affecting the diffuse contribution
-color (range <0; 1>). Reflection to diffuse proportion is set with reflectivity_index
-and reflectivity_range (both in range <0; 1>), where:
-
-- (1, 1) results with a mirror-like appearance
-- (0, 0) results with a diffuse-like appearance
-- intermediate values result with a plastic-like appearance and various gloss profiles. 
+Strongly reflective, metalic material. Note, this material has default values: ``reflectivity_index = 1``
+and ``reflectivity_range = 1``. In this configuration the shading algorithm overrides ``surface_albedo``
+with the color assigned to each primitive (RGB range <0; 1>), which results with colorized reflections.
+Roughness of the surface should be usually small.
 """
 
 m_plastic = {
@@ -90,6 +92,24 @@ set above 0 (up to 1, resulting with mirror-like appearance) or minimized with a
 reflectivity_range value (down to 0). Higher refraction_index gives a more glossy look.
 """
 
+m_matt_plastic = {
+      "ClosestHitPrograms": ["0::path_tracing_materials.ptx::reflective_closest_hit"],
+      "AnyHitPrograms": ["1::path_tracing_materials.ptx::any_hit"],
+      "VarInt": { "material_flags": 6 },
+      "VarFloat": {
+        "reflectivity_index": 0.0,
+        "reflectivity_range": 0.5,
+        "surface_roughness": 0.001
+      },
+      "VarFloat3": {
+        "refraction_index": [ 2.0, 2.0, 2.0 ],
+        "surface_albedo": [ 1.0, 1.0, 1.0 ]
+      }
+}
+"""
+Similar to :attr:`plotoptix.materials.m_plastic` but slightly rough surface.
+"""
+
 m_clear_glass = {
       "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::glass_closest_hit" ],
       "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
@@ -106,7 +126,27 @@ m_clear_glass = {
     }
 """
 Glass, with reflection and refraction simulated. Color components meaning is "attenuation length"
-and the range is <0; inf>.
+and the color range is <0; inf>.
+"""
+
+m_matt_glass = {
+      "ClosestHitPrograms": [ "0::path_tracing_materials.ptx::glass_closest_hit" ],
+      "AnyHitPrograms": [ "1::path_tracing_materials.ptx::any_hit" ],
+      "VarInt": { "material_flags": 12 },
+      "VarFloat": {
+        "radiation_length": 0.0,
+        "vol_scattering": 1.0,
+        "light_emission": 0.0,
+        "surface_roughness": 0.2
+      },
+      "VarFloat3": {
+        "refraction_index": [ 1.4, 1.4, 1.4 ],
+        "surface_albedo": [ 1.0, 1.0, 1.0 ]
+      }
+    }
+"""
+Glass with surface roughness configured to obtain matt appearance. Color components meaning is
+"attenuation length" and the color range is <0; inf>.
 """
 
 m_dispersive_glass = {
@@ -124,7 +164,7 @@ m_dispersive_glass = {
       }
     }
 """
-Glass, with reflection and refraction simulated. Refraction index is varying with the wavelength,
+Clear glass, with reflection and refraction simulated. Refraction index is varying with the wavelength,
 resulting with the light dispersion. Color components meaning is "attenuation length"
 and the range is <0; inf>.
 """